impl: using transmute
