+++ crates/prover/src/core/backend/web/fri.rs
@@ -1,0 +1,34 @@
+use super::WebBackend;
+use crate::core::fields::qm31::SecureField;
+use crate::core::fri::FriOps;
+use crate::core::poly::circle::SecureEvaluation;
+use crate::core::poly::line::LineEvaluation;
+use crate::core::poly::twiddles::TwiddleTree;
+use crate::core::poly::BitReversedOrder;
+
+// TODO(andrew) Is this optimized?
+#[allow(unused_variables)]
+impl FriOps for WebBackend {
+    fn fold_line(
+        eval: &LineEvaluation<Self>,
+        alpha: SecureField,
+        twiddles: &TwiddleTree<Self>,
+    ) -> LineEvaluation<Self> {
+        todo!()
+    }
+
+    fn fold_circle_into_line(
+        dst: &mut LineEvaluation<Self>,
+        src: &SecureEvaluation<Self, BitReversedOrder>,
+        alpha: SecureField,
+        twiddles: &TwiddleTree<Self>,
+    ) {
+        todo!()
+    }
+
+    fn decompose(
+        eval: &SecureEvaluation<Self, BitReversedOrder>,
+    ) -> (SecureEvaluation<Self, BitReversedOrder>, SecureField) {
+        todo!()
+    }
+}
